Moved the readAndParseFile to helper file, so it does not close the open connection and cause the test to fail
